+++ openrouteservice/src/main/java/org/heigit/ors/routing/pathprocessors/BordersExtractor.java
@@ -1,7 +1,10 @@
 package org.heigit.ors.routing.pathprocessors;
 
+import com.carrotsearch.hppc.IntHashSet;
 import org.heigit.ors.routing.graphhopper.extensions.storages.BordersGraphStorage;
 
+import java.util.ArrayList;
+import java.util.List;
 
 public class BordersExtractor {
     public enum Avoid { CONTROLLED, NONE, ALL }
@@ -41,5 +44,28 @@ public class BordersExtractor {
             }
         }
         return false;
+    }
+
+    /**
+     * Check whether the start and end nodes of a list of edges are in the same country.
+     * @param edgeIds Edges that the country should be checked for
+     * @return true if at least one node is in the same country
+     */
+    public boolean isSameCountry(List<Integer> edgeIds){
+        if(edgeIds.isEmpty())
+            return true;
+
+        short country0 = storage.getEdgeValue(edgeIds.get(0), BordersGraphStorage.Property.START);
+        short country1 = storage.getEdgeValue(edgeIds.get(0), BordersGraphStorage.Property.END);
+        for(int edgeId : edgeIds) {
+            short country2 = storage.getEdgeValue(edgeId, BordersGraphStorage.Property.START);
+            short country3 = storage.getEdgeValue(edgeId, BordersGraphStorage.Property.END);
+            if(country0 != country2
+            && country0 != country3
+            && country1 != country2
+            && country1 != country3)
+                return false;
+        }
+        return true;
     }
 }