+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
+- Check whether routing points are within different countries before routing and break if they are and all borders should be avoided
 ### Fixed
+- Updated Docker process to use Java 11 ([#777](https://github.com/GIScience/openrouteservice/issues/777))
 ### Changed
 ### Deprecated
 
@@ -22,6 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Improve recommended weighting for cycling and walking profiles ([#665](https://github.com/GIScience/openrouteservice/issues/665))
 - Restructure AdditionWeighting
+- Upgrade to Java 11
 ### Deprecated
 - Use recommended weighting instead of fastest ([#763](https://github.com/GIScience/openrouteservice/issues/763))